+++ ChangeLog.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [un-released] - 2020-06-06
+
+### Added
+
+- Feat: add two options for TLS connection to the broker
+- Feat: add support to Certificate Authorities
+
 ## [0.1.1] - 2020-06-06
 
 ### Added